requirements-pyfuse3.txt
# Lighthouse Bridge Requirements - pyfuse3 version
# Complete package list including optional packages

# Core FUSE support - both for compatibility during transition
fusepy==3.0.1  # Current Bridge uses this
pyfuse3==3.4.0  # Future version
trio==0.26.0  # Required by pyfuse3 for async

# Core async dependencies
aiofiles==24.1.0
aiohttp==3.9.5
aiosqlite==0.20.0

# Web framework for API
fastapi>=0.115.12
uvicorn[standard]>=0.30.1
pydantic>=2.8.0

# Security
cryptography==42.0.8

# Serialization
msgpack==1.0.8
orjson==3.10.5  # Fast JSON

# System monitoring
psutil==5.9.8

# Utilities
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1  # Better console output

# Bridge server dependencies (MCP stays on host)

# Optional but recommended packages

# Machine Learning (for pattern cache optimization)
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.14.0

# Code analysis (for AST operations)
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0
tree-sitter-rust==0.21.2
tree-sitter-go==0.21.0

# Database drivers
asyncpg==0.29.0  # PostgreSQL
aiomysql==0.2.0  # MySQL
redis==5.0.6  # Redis
motor==3.4.0  # MongoDB

# Monitoring and observability
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0

# Structured logging
structlog==24.2.0

# Development and testing tools (consider removing for production)
# pytest==8.2.2
# pytest-asyncio==0.23.7
# pytest-cov==5.0.0
# black==24.4.2
# mypy==1.10.0
# ruff==0.4.9
# ipython==8.25.0  # Interactive debugging
# ipdb==0.13.13  # Python debugger

# Performance profiling
py-spy==0.3.14
memory-profiler==0.61.0
line-profiler==4.1.2

# Additional utilities
httpx>=0.27.1  # HTTP client
tenacity==8.3.0  # Retry logic
cachetools==5.3.3  # Caching utilities
watchdog==4.0.1  # File system monitoring